Generate front matter programatically
Infer when assertion transformations introduce dependencies on external
scripts in the Test262 harness.
